// src/types/clinic.d.ts
	interface IFollowUpRecord {
		_id: string;
		type: 'FOLLOWUP';
		uploadedBy: 'CLINIC';
		url: string | null;
		comment: string | null;
		vaccinationDate: string | null;
		nextVaccinationDate: string | null;
		vaccinatedOnDate: string | null;
		followUpDate: string;
		followUpType: string;
		vaccineName: string | null;
		notificationCount: number;
		active: boolean;
		parent: ICommonTypes.IParent;
		clinic: ICommonTypes.IClinic;
		pet: ICommonTypes.IPet;
		createdAtIST: string;
		updatedAtIST: string;
		recordId: string;
		createdAtUTC: string;
		updatedAtUTC: string;
		followUpCompleteDate: string;
	}
// ...
	interface IParent {
		name: string;
		mobile: string;
		parentId: string;
	}